Create clue_removal module
